+++ backend/src/controller/timeslot.rs
@@ -1,8 +1,9 @@
 use crate::service::timeslot::TimeslotService;
+ use crate::domain::organisation::OrgID;
 use chrono::NaiveDate;
 use crate::parser::*;
-use crate::domain::User;
-use crate::domain::Day;
+use crate::domain::user::User;
+use crate::domain::timeslot::Day;
 use rocket::post;
 use rocket::response::status;
 use rocket::routes;
@@ -42,7 +43,7 @@ async fn subscirbe_to_hours(
     user: User,
     payload: Json<SubscribeToHoursRequest>,
 ) -> Json<bool> {
-    let org_id = Uuid::parse_str(&payload.org_id).expect("Faild to parse org_id");
+    let org_id = payload.org_id.clone();
     let date: NaiveDate = parse_iso8601_date(&payload.date).expect("Failed to parse date");
     let user_id = user.user_id;
     let hours = payload.hours.clone();
@@ -80,7 +81,7 @@ async fn get_days(
     user: User,
     payload: Json<GetDaysRequest>,
 ) -> Result<Json<Vec<Day>>, status::Custom<String>> {
-    let org_id = Uuid::parse_str(&payload.org_id).expect("faild to parse org_id");
+    let org_id = payload.org_id.clone();
     let start_date: NaiveDate = parse_iso8601_date(&payload.start_date).expect("Failed to parse start date");
     let end_date: NaiveDate = parse_iso8601_date(&payload.end_date).expect("Failed to parse end date");
