+++ backend/src/controller/user.rs
@@ -1,4 +1,4 @@
-use crate::domain::User;
+use crate::domain::user::*;
 use rocket::put;
 use crate::service::user::UserService;
 use rocket::get;
@@ -11,7 +11,6 @@ use serde::Deserialize;
 use serde::Serialize;
 use std::sync::Arc;
 use utoipa::ToSchema;
-use uuid::Uuid;
 
 /// Request body for creating a user.
 #[derive(Debug, Serialize, Deserialize, ToSchema)]
