+++ backend/src/domain/mod.rs
@@ -1,0 +1,4 @@
+pub mod user;
+pub mod organisation;
+pub mod timeslot;
+pub mod access_notification;
